+++ seantis/agencies/types/__init__.py
@@ -73,6 +73,7 @@ class IOrganization(form.Schema):
                 SimpleTerm(value=u'role', title=_('Role')),
                 SimpleTerm(value=u'title', title=_('Last Name First Name')),
                 SimpleTerm(value=u'start', title=_('Start of membership')),
+                SimpleTerm(value=u'postfix', title=_('Postfix')),
                 SimpleTerm(value=u'lastname', title=_('Last Name')),
                 SimpleTerm(value=u'firstname', title=_('First Name')),
                 SimpleTerm(value=u'year', title=_('Year')),
@@ -127,6 +128,11 @@ class IMembership(IPeopleMembership):
 
     prefix = schema.TextLine(
         title=_(u"Prefix"),
+        required=False
+    )
+
+    postfix = schema.TextLine(
+        title=_(u"Postfix"),
         required=False
     )